+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Assert-ObjectProperty` (alias `Should-HaveProperty`) - Asserts that an
     object contains a specified property and optionally that the property has
     a specified value.
+  - `Assert-ObjectMethod` (alias `Should-HaveMethod`) - Asserts that an object
+    contains a specified method.
 
 ### Changed